[jsk_pcl_ros, jsk_perception] Move mask image operation to jsk_perception
