+++ jsk_pcl_ros/README.md
@@ -178,24 +178,6 @@ add two different `pcl_msgs/PointIndices` into one indices.
   If this parameter is true, `~input/src1` and `~input/src2` are synchronized with
   approximate time policy.
 
-### jsk\_pcl/PolygonToMaskImage
-![](images/polygon_to_mask_image.png)
-
-Convert polygon into mask image.
-
-#### Subscribing Topic
-* `~input` (`geometry_msgs/PolygonStamped`)
-
-  Input 3-D polygon.
-* `~input/camera_info` (`sensor_msgs/CameraInfo`)
-
-  Input camera info to project 3-D polygon.
-
-#### Publishing Topic
-* `~output` (`sensor_msgs/Image`)
-
-  Mask image to fill `~input` polygon. Currently only convex polygon is supported.
-
 ### jsk\_pcl/MaskImageFilter
 ![](images/mask_image_filter.png)
 
@@ -217,62 +199,6 @@ Extract indices of pointcloud which is masked by mask image. The pointcloud is n
 
   Indices of the points masked by `~input/mask`.
 
-### jsk\_pcl/ROIToRect
-Convert camera info with ROI to `geometry_msgs/PolygonStamped`.
-
-#### Subscribing Topic
-* `~input` (`sensor_msgs/CameraInfo`)
-
-  Input camera info with ROI filled.
-
-#### Publishing Topic
-* `~output` (`geometry_msgs/PolygonStamped`)
-
-  Output rectangle region.
-
-### jsk\_pcl/ROIToMaskImage
-Convert camera info with ROI to mask image.
-
-#### Subscribing Topic
-* `~input` (`sensor_msgs/CameraInfo`)
-
-  Input camera info with ROI filled.
-
-#### Publishing Topic
-* `~output` (`sensor_msgs/Image`)
-
-  Output mask image.
-
-### jsk\_pcl/MaskImageToROI
-Convert a mask image into camera info with roi.
-
-#### Subscribing Topic
-* `~input` (`sensor_msgs/Image`)
-
-  Input mask image.
-* `~input/camera_info` (`sensor_msgs/CameraInfo`)
-
-  Original camera info.
-
-#### Publishing Topic
-* `~output` (`sensor_msgs/CameraInfo`)
-
-  Camera info with ROI field filled.
-
-### jsk\_pcl/MaskImageToRect
-Convert a mask image into geometry_msgs::PolygonStamped.
-
-#### Subscribing Topic
-* `~input` (`sensor_msgs/Image`)
-
-  Input mask image.
-
-#### Publishing Topic
-* `~output` (`geometry_msgs/PolygonStamped`)
-
-  PolygonStamped message which only contains two points. Minimum point and Maximum point to represent bounding box in image.
-
-### jsk\_pcl/MaskImageToRect
 ### jsk\_pcl/TorusFInder
 ![](images/torus_finder.png)
 
@@ -311,41 +237,6 @@ Find a torus out of pointcloud based on RANSAC with 3-D circle model.
 * `~max_iterations` (Integer, default: `100`)
 
   Maximum number of iterations of RANSAC.
-### jsk\_pcl/RectToROI
-Convert rectangle (`geometry_msgs/Polygon`) into ROI with camera info (`sensor_msgs/CameraInfo`).
-
-We expect it will be used with image_view2.
-
-#### Subscribing Topic
-* `~input` (`geometry_msgs/Polygon`)
-
-  Polygon to represent rectangle region of image.
-* `~input/camera_info` (`sensor_msgs/CameraInfo`)
-
-  Original camera info.
-
-#### Publishing Topic
-* `~output` (`sensor_msgs/CameraInfo`)
-
-  camera info with ROI filled by `~input`.
-
-### jsk\_pcl/RectToMaskImage
-Convert rectangle (`geometry_msgs/Polygon`) into mask image (`sensor_msgs/Image`)
-
-We expect it will be used with image_view2.
-
-#### Subscribing Topic
-* `~input` (`geometry_msgs/Polygon`)
-
-  Polygon to represent rectangle region of image.
-* `~input/camera_info` (`sensor_msgs/CameraInfo`)
-
-  Original camera info.
-
-#### Publishing Topic
-* `~output` (`sensor_msgs/Image`)
-
-  Mask image.
 
 ### jsk\_pcl/AddColorFromImage
 ![](images/add_color_from_image.png)
