+++ jsk_perception/images/polygon_to_mask_image.png
